workout backup sortu

+++ src/model/Workout.java
@@ -3,11 +3,11 @@ package model;
 public class Workout {
 	
 	private String izena;
-	private double maila;
+	private int maila;
 	private String video_url;
 	private int ariketaSize;
 	
-	public Workout(String name, double maila, String video_url, int ariketaSize) {
+	public Workout(String name, int maila, String video_url, int ariketaSize) {
 		this.izena = name;
 		this.maila = maila;
 		this.video_url = video_url;
@@ -22,11 +22,11 @@ public class Workout {
 		this.izena = izena;
 	}
 
-	public double getMaila() {
+	public int getMaila() {
 		return maila;
 	}
 
-	public void setMaila(double maila) {
+	public void setMaila(int maila) {
 		this.maila = maila;
 	}
